Add training script for ICL experiment.
Add an example training script to reproduce the in-context learning experiment from the paper (see issue #4).

An important detail is to set `--unembed_mask 0` (otherwise the model will be prevented from predicting the `unk`, which is used for this task). You may need to run the script with multiple seeds (e.g. 10) to get an initialization that learns to solve the task.